+ Static Protobufs
+ fixed localPlayer acquisition mechanism
+ fixed rune pickup not recognizing the rune and its type
+ added hook on/off setters to prevent unnesecessary desync
+ texture unloading on quitting game

+++ Dota2Cheat/Hooks/RunFrame.cpp
@@ -84,7 +84,7 @@ void Hooks::UpdateWeather() {
 	varInfo.var->value.i32 = Config::WeatherListIdx;
 }
 
-void Hooks::hkRunFrame(u64 a, u64 b) {
+void Hooks::hkRunFrame(uintptr_t a, uintptr_t b) {
 	bool isInGame = Interfaces::Engine->IsInGame();
 
 	if (isInGame) {
